#5 [Chore] 스타일 컴포넌트 분리
